Cree muchos factorys para muchas cosas owowo

+++ src/model/items/Axe.java
@@ -20,7 +20,7 @@ public class Axe extends AbstractItem {
    * @param minRange  the minimum range of the Axe.
    * @param maxRange  the maximum range of the Axe.
    */
-  public Axe(final String name, final int power, final int minRange, final int maxRange) {
+  public Axe(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
   }
