+++ src/model/items/Bow.java
@@ -22,7 +22,7 @@ public class Bow extends AbstractItem {
    * @param minRange  the minimum range of the Bow.
    * @param maxRange  the maximum range of the Bow.
    */
-  public Bow(final String name, final int power, final int minRange, final int maxRange) {
+  public Bow(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
     this.minRange = Math.max(minRange, 2);
     this.maxRange = Math.max(maxRange, this.minRange);
