+++ src/model/items/Spear.java
@@ -20,7 +20,7 @@ public class Spear extends AbstractItem {
    * @param minRange the minimum range of the Spear.
    * @param maxRange the maximum range of the Spear.
    */
-  public Spear(final String name, final int power, final int minRange, final int maxRange) {
+  public Spear(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
   }
