+++ src/model/items/Staff.java
@@ -20,7 +20,7 @@ public class Staff extends AbstractItem {
    * @param minRange  the minimum range of the Staff.
    * @param maxRange  the maximum range of the Staff.
    */
-  public Staff(final String name, final int power, final int minRange, final int maxRange) {
+  public Staff(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
   }
