+++ src/model/items/Sword.java
@@ -20,7 +20,7 @@ public class Sword extends AbstractItem {
    * @param minRange  the minimum range of the Sword.
    * @param maxRange  the maximum range of the Sword.
    */
-  public Sword(final String name, final int power, final int minRange, final int maxRange) {
+  public Sword(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
   }
